refactor(settings): retire duplicate history surface
