+++ packages/desktop/tests/e2e/workspaceTabs.spec.ts
@@ -493,7 +493,7 @@ test('loads tools, centers every compact nav icon, and customizes shortcuts', as
     .toBeLessThanOrEqual(520);
 
   await page.evaluate(() => {
-    window.postMessage({ command: 'setTab', tabIndex: 5 }, '*');
+    window.postMessage({ command: 'setTab', tabIndex: 4 }, '*');
   });
   await expect(page.locator('tools-tab tool-card').first()).toBeVisible({
     timeout: 5_000,
@@ -531,7 +531,7 @@ test('loads tools, centers every compact nav icon, and customizes shortcuts', as
   }
 
   await page.evaluate(() => {
-    window.postMessage({ command: 'setTab', tabIndex: 11 }, '*');
+    window.postMessage({ command: 'setTab', tabIndex: 10 }, '*');
   });
   const shortcuts = page.locator('shortcuts-tab');
   await expect(
