+++ src/test-kernel/desktop/DesktopCommandSurface.vitest.ts
@@ -36,7 +36,6 @@ const DESKTOP_DISPATCH_CASES: ReadonlyArray<
   ['texra.desktop.toggleSidePanel', 'toggleSidePanel', []],
   ['texra.desktop.toggleSummaryBar', 'toggleSummaryBar', []],
   ['texra.showMemory', 'showSettings', [SETTINGS_TAB.MEMORY]],
-  ['texra.showAgentHistory', 'showSettings', [SETTINGS_TAB.HISTORY]],
   ['texra.showModels', 'showSettings', [SETTINGS_TAB.MODELS]],
   ['texra.showAgents', 'showSettings', [SETTINGS_TAB.AGENTS]],
   ['texra.showTools', 'showSettings', [SETTINGS_TAB.TOOLS]],
@@ -254,7 +253,6 @@ describe('desktop command surface', () => {
       'New',
       'separator',
       'Show Memory',
-      'Show Agent Execution History',
       'Show Models',
       'Show Agents',
       'Show Tool Dashboard',
