+++ src/test-kernel/desktop/loadSourceModule.ts
@@ -31,7 +31,6 @@ interface TestSourceModules {
   '@desktop/main/desktopExecutionIpc': typeof import('@desktop/main/desktopExecutionIpc');
   '@desktop/main/desktopFileSelection': typeof import('@desktop/main/desktopFileSelection');
   '@desktop/main/desktopMenuTemplate': typeof import('@desktop/main/desktopMenuTemplate');
-  '@desktop/main/desktopHistoryHandlers': typeof import('@desktop/main/desktopHistoryHandlers');
   '@desktop/main/desktopLogIpc': typeof import('@desktop/main/desktopLogIpc');
   '@desktop/main/desktopNavigationPolicy': typeof import('@desktop/main/desktopNavigationPolicy');
   '@desktop/main/desktopOnboardingIpc': typeof import('@desktop/main/desktopOnboardingIpc');
